requirements.txt
openai
python-dotenv
sentence-transformers
transformers
modal
joblib
pandas
beautifulsoup4
requests
tqdm
feedparser
pydantic
chromadb
